PlantGen-API/requirements.txt
# FastAPI y servidor ASGI
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# MongoDB async driver
motor==3.3.2
pymongo==4.6.1

# Variables de entorno
python-dotenv==1.0.1

# CORS y seguridad
python-multipart==0.0.9

# Utilidades
typing-extensions==4.9.0
